+++ src/shapeandshare/dicebox/connectors/filesystem_connector.py
@@ -1,16 +1,20 @@
-import array
 import fnmatch
 import logging
 import os
 import struct
-from typing import Dict, Any, List, Union
+from typing import Dict, Any, List, Union, Optional
 
 import numpy
 from PIL import Image
-from numpy import ndarray
+from numpy import ndarray, int32, float32
 
 from ..config.dicebox_config import DiceboxConfig
 from ..utils.helpers import lucky
+from tensorflow.keras.preprocessing.image import img_to_array
+
+# can be used as a foreign (sane) data set for validation of structures.
+# from tensorflow import shape
+# from tensorflow.keras.datasets import cifar10
 
 
 class FileSystemConnector:
@@ -40,6 +44,48 @@ class FileSystemConnector:
         else:
             logging.info('File System Connector Data Indexing Disabled.')
 
+    def get_data_set(self) -> dict:
+        """
+        Returns a dictionary of [k:filename, v:array of filename and category] for the entire data set.
+
+        When we move this to python 3, there are much better libraries to handle this.  Checkout PurePath..
+        """
+        data_set: Dict[str, List[Union[str, Dict[str, Any]]]] = {}
+        category: Dict[str, Any]
+        # filename: str
+        for root, _, filenames in os.walk(self.data_directory):
+            for filename in fnmatch.filter(filenames, '*.png'):
+                new_entry = str(os.path.join(root, filename))
+                new_entry = new_entry.replace('%s%s' % (self.data_directory, os.path.sep), '')
+                new_entry = os.path.normpath(new_entry)
+                category, filename = new_entry.split(os.path.sep)
+                # logging.info("category: (%s), filename: (%s)" % (category, filename))
+                data_set[new_entry]: Union[str, Dict[str, Any]] = [filename, category]
+        # logging.info(data_set)
+        return data_set
+
+    def get_data_set_categories(self) -> Dict[Any, int]:
+        """Returns the dataset index as a dictionary of categories
+        :return: category map as a dictionary
+        """
+        natural_categories = []
+        category_map = {}
+        value_list = self.dataset_index.values()
+        # print(self.dataset_index)
+        # print(value_list)
+        for item in value_list:
+            # logging.info(item)
+            # logging.info("natural category label: (%s)" % item[1])
+            natural_categories.append(item[1])
+        natural_categories = sorted(set(natural_categories))
+        cat_index = 0
+        for cat in natural_categories:
+            # logging.info("%i: %s" % (cat_index, cat))
+            category_map[cat] = cat_index
+            cat_index += 1
+        # logging.info(category_map)
+        return category_map
+
     def get_batch_list(self, batch_size: int) -> List[int]:
         """For a given batch size, returns a random selection of indices
 
@@ -66,18 +112,18 @@ class FileSystemConnector:
             output.append(value_list[i])
         return output
 
-    def get_batch(self, batch_size: int, noise: float = 0.0) -> List[Union[list, List[ndarray]]]:
+    def get_batch(self, batch_size: int, noise: float = 0.0) -> List[Union[ndarray, ndarray]]:
         """For a given batch size and noise level, returns a dictionary of data and labels.
 
         :param batch_size: integer
         :param noise: floating point number
         :return: dictionary of image data and labels
         """
-        image_data = []
-        image_labels = []
+        image_data: ndarray = numpy.empty((0, self.config.IMAGE_WIDTH, self.config.IMAGE_HEIGHT, 3), dtype=float32)
+        image_labels: ndarray = numpy.empty((0, 1), dtype=int32)
 
         category_map: Dict[Any, int] = self.category_map
-        batch_list: [] = self.get_batch_list(batch_size)
+        batch_list: List[int] = self.get_batch_list(batch_size)
 
         # build file path
         for i in range(0, batch_size):
@@ -85,20 +131,18 @@ class FileSystemConnector:
             filename = "%s/%s/%s" % (self.data_directory, item[1], item[0])
             logging.debug("(%s)(%s)", category_map[item[1]], filename)
             # logging.info("  natural category label: (%s)" % item[1])
-            # logging.info("  neural __network category label: (%i)" % category_map[item[1]])
-            cat_one_hot = numpy.zeros(len(category_map))
-            cat_one_hot[int(category_map[item[1]])] = 1
-            image_labels.append(cat_one_hot)
-            # image_labels.append(category_map[item[1]])
+            # print("  neural network category label: (%i)" % category_map[item[1]])
 
+            image_labels = numpy.append(image_labels, numpy.array([[category_map[item[1]]]], dtype=int32), axis=0)
+
+            ### Cache Invalidation, hahahaha!
             # Help prevent over-fitting, and allow for new
             # sensory data to enter the cache, even when a cache
             # hit would occur.
             if filename in self.pixel_cache and not lucky(noise):
                 del self.pixel_cache[filename]
 
-            # use pixel cache if possible
-            # [k,v] (filename, pixeldata)
+            # use pixel cache if possible, [k,v] (filename, pixeldata)
             if filename in self.pixel_cache:
                 # found in cache
                 pixel_data = self.pixel_cache[filename]
@@ -108,19 +152,25 @@ class FileSystemConnector:
                 self.pixel_cache[filename] = pixel_data  # add to cache
                 logging.debug("cached pixel data for (%s)", filename)
 
-            image_data.append(pixel_data)
+            # add new image to the set
+            image_data = numpy.append(image_data, numpy.array([pixel_data], dtype=float32), axis=0)
+
+        # Validation using another data set , note the dimensions are 32x32
+        # (input_train, target_train), (input_test, target_test) = cifar10.load_data()
+        # Parse numbers as floats
+        # input_train = input_train.astype('float32')
+        # input_test = input_test.astype('float32')
+        # Scale data
+        # input_train = input_train / 255
+        # input_test = input_test / 255
+        # return [input_train, target_train]
+
         return [image_data, image_labels]
 
-    def process_image(self, filename: str, noise: float = 0.0):
-        """For a given filename, and noise level, will return a numpy array of pixel data.
+    def process_image(self, filename: str, noise: float = 0.0) -> Optional[ndarray]:
+        # For a given filename, and noise level, will return a numpy array of pixel data.
 
-        :param filename:
-        :param noise:
-        :return:
-        """
-        pixel_data = array.array('B')
-
-        local_image = Image.open(filename).convert('L')  # Load as gray
+        local_image = Image.open(filename).convert('RGB')  # Load as RGB
 
         original_width, original_height = local_image.size
         # original_size = original_width, original_height
@@ -183,63 +233,6 @@ class FileSystemConnector:
         # filename = datetime.now().strftime('transform_%Y-%m-%d_%H_%M_%S_%f.png')
         # local_image.save("./tmp/%s" % filename)
 
-        pixel = local_image.load()
-
-        width, height = local_image.size
-
-        for x in range(0, width):
-            for y in range(0, height):
-                pixel_value = pixel[x, y]
-                # logging.info("(%f)" % float(noise))
-                # chance = float(ord(struct.unpack('c', os.urandom(1))[0])) / 255
-                # logging.info("(%f)" % chance)
-                # if float(noise) >= float(chance):
-                #  # logging.info("  adding pixel noise (%f/%f)" % (float(noise), chance)) # add noise
-                #  pixel_value = int(ord(struct.unpack('c', os.urandom(1))[0]))
-                pixel_data.append(pixel_value)
-
-        data = numpy.frombuffer(pixel_data, dtype=numpy.uint8)
-        return data
-
-    def get_data_set_categories(self) -> Dict[Any, int]:
-        """Returns the dataset index as a dictionary of categories
-
-        :return: category map as a dictionary
-        """
-        natural_categories = []
-        category_map = {}
-        value_list = self.dataset_index.values()
-        # print(self.dataset_index)
-        # print(value_list)
-        for item in value_list:
-            # logging.info(item)
-            # logging.info("natural category label: (%s)" % item[1])
-            natural_categories.append(item[1])
-        natural_categories = sorted(set(natural_categories))
-        cat_index = 0
-        for cat in natural_categories:
-            # logging.info("%i: %s" % (cat_index, cat))
-            category_map[cat] = cat_index
-            cat_index += 1
-        # logging.info(category_map)
-        return category_map
-
-    def get_data_set(self) -> dict:
-        """
-        Returns a dictionary of [k:filename, v:array of filename and category] for the entire data set.
-
-        When we move this to python 3, there are much better libraries to handle this.  Checkout PurePath..
-        """
-        data_set: Dict[str, List[Union[str, Dict[str, Any]]]] = {}
-        category: Dict[str, Any]
-        # filename: str
-        for root, _, filenames in os.walk(self.data_directory):
-            for filename in fnmatch.filter(filenames, '*.png'):
-                new_entry = str(os.path.join(root, filename))
-                new_entry = new_entry.replace('%s%s' % (self.data_directory, os.path.sep), '')
-                new_entry = os.path.normpath(new_entry)
-                category, filename = new_entry.split(os.path.sep)
-                # logging.info("category: (%s), filename: (%s)" % (category, filename))
-                data_set[new_entry]: Union[str, Dict[str, Any]] = [filename, category]
-        # logging.info(data_set)
-        return data_set
+        # https://www.tensorflow.org/api_docs/python/tf/keras/preprocessing/image/img_to_array
+        image_data = img_to_array(local_image)
+        return image_data