+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+3.4.0
+-----
+- RGB Support
+- Additional activation functions, and optimizers
+- Added Conv2D layer support
+
 3.3.0
 -----
 - Pathing fixes for Windows